Move worker to its own directory(part of worker code refactoring)
